arweave_uploader_service/pyproject.toml
[tool.poetry]
name = "arweave-uploader-service"
version = "0.1.0"
description = "Arweave uploader microservice for chatbot"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
arweave-python-client = "^1.0.19"
python-multipart = "^0.0.6"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
